Fix for bug 41748: Significantly reduced SQL overhead when moving files around file storages

+++ com.openexchange.admin/src/com/openexchange/admin/tools/filestore/MasterUser2UserFilestoreDataMover.java
@@ -49,11 +49,12 @@
 
 package com.openexchange.admin.tools.filestore;
 
+import static com.openexchange.filestore.FileStorages.ensureEndingSlash;
+import static com.openexchange.filestore.FileStorages.getFullyQualifyingUriForUser;
 import static com.openexchange.filestore.FileStorages.getQuotaFileStorageService;
 import java.io.IOException;
 import java.net.URI;
 import java.sql.SQLException;
-import java.util.Map;
 import java.util.Set;
 import org.slf4j.Logger;
 import com.openexchange.admin.rmi.dataobjects.Context;
@@ -128,12 +129,21 @@ public class MasterUser2UserFilestoreDataMover extends FilestoreDataMover {
             Set<String> srcFiles = determineFileLocationsFor(dstUserId, contextId);
             if (false == srcFiles.isEmpty()) {
                 // Copy each file from source to destination
-                Map<String, String> prevFileName2newFileName = copyFiles(srcFiles, srcStorage, dstStorage);
+                URI srcFullUri = ensureEndingSlash(getFullyQualifyingUriForUser(masterUserId, contextId, srcBaseUri));
+                URI dstFullUri = ensureEndingSlash(getFullyQualifyingUriForUser(dstUserId, contextId, dstBaseUri));
+                CopyResult copyResult = copyFiles(srcFiles, srcStorage, dstStorage, srcFullUri, dstFullUri);
 
-                // Propagate new file locations throughout registered FilestoreLocationUpdater instances
-                propagateNewLocations(prevFileName2newFileName);
+                if (Operation.COPIED.equals(copyResult.operation)) {
+                    // Propagate new file locations throughout registered FilestoreLocationUpdater instances
+                    propagateNewLocations(copyResult.prevFileName2newFileName);
 
-                srcStorage.deleteFiles(srcFiles.toArray(new String[srcFiles.size()]));
+                    srcStorage.deleteFiles(srcFiles.toArray(new String[srcFiles.size()]));
+                } else {
+                    // Simply adjust filestore_usage entries
+                    long totalSize = copyResult.totalSize;
+                    decUsage(totalSize, masterUserId, contextId);
+                    incUsage(totalSize, dstUserId, contextId);
+                }
             }
         } catch (OXException e) {
             throw new StorageException(e);
