+++ com.openexchange.admin/src/com/openexchange/admin/tools/filestore/User2MasterUserFilestoreDataMover.java
@@ -49,12 +49,13 @@
 
 package com.openexchange.admin.tools.filestore;
 
+import static com.openexchange.filestore.FileStorages.ensureEndingSlash;
+import static com.openexchange.filestore.FileStorages.getFullyQualifyingUriForUser;
 import static com.openexchange.filestore.FileStorages.getQuotaFileStorageService;
 import java.io.File;
 import java.io.IOException;
 import java.net.URI;
 import java.sql.SQLException;
-import java.util.Map;
 import java.util.Set;
 import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
@@ -114,6 +115,9 @@ public class User2MasterUserFilestoreDataMover extends FilestoreDataMover {
         int masterUserId = masterUser.getId().intValue();
         try {
             // Check
+            if (masterUser.getFilestoreOwner().intValue() <= 0) {
+                throw new StorageException("Master user " + masterUserId + " has no individual file storage.");
+            }
             if (srcUser.getFilestoreOwner().intValue() > 0 && srcUser.getFilestoreOwner().intValue() != srcUserId) {
                 throw new StorageException("User's file storage does not belong to user. Owner " + srcUser.getFilestoreOwner() + " is not equal to " + srcUser.getId());
             }
@@ -125,12 +129,21 @@ public class User2MasterUserFilestoreDataMover extends FilestoreDataMover {
             // Copy each file from source to destination
             Set<String> srcFiles = srcStorage.getFileList();
             if (false == srcFiles.isEmpty()) {
-                Map<String, String> prevFileName2newFileName = copyFiles(srcFiles, srcStorage, dstStorage);
+                URI srcFullUri = ensureEndingSlash(getFullyQualifyingUriForUser(srcUserId, contextId, srcBaseUri));
+                URI dstFullUri = ensureEndingSlash(getFullyQualifyingUriForUser(masterUserId, contextId, dstBaseUri));
+                CopyResult copyResult = copyFiles(srcFiles, srcStorage, dstStorage, srcFullUri, dstFullUri);
 
-                // Propagate new file locations throughout registered FilestoreLocationUpdater instances
-                propagateNewLocations(prevFileName2newFileName);
+                if (Operation.COPIED.equals(copyResult.operation)) {
+                    // Propagate new file locations throughout registered FilestoreLocationUpdater instances
+                    propagateNewLocations(copyResult.prevFileName2newFileName);
 
-                srcStorage.deleteFiles(srcFiles.toArray(new String[srcFiles.size()]));
+                    srcStorage.deleteFiles(srcFiles.toArray(new String[srcFiles.size()]));
+                } else {
+                    // Simply adjust filestore_usage entries
+                    long totalSize = copyResult.totalSize;
+                    decUsage(totalSize, srcUserId, contextId);
+                    incUsage(totalSize, masterUserId, contextId);
+                }
             }
 
             if ("file".equalsIgnoreCase(srcBaseUri.getScheme())) {
